serverutils,server: remove dependency from testutils/serverutils to kv
... by moving a function from serverutils and putting it in sqlutils
(which is a less important package which can depende on random crap).

Release note: None
