+++ pkg/sql/create_stats_test.go
@@ -19,6 +19,7 @@ import (
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/base"
+	"github.com/cockroachdb/cockroach/pkg/kv"
 	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/sql/distsqlrun"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
@@ -97,7 +98,9 @@ func TestStatsWithLowTTL(t *testing.T) {
 				return
 			}
 			// Force a table GC of values older than 1 second.
-			serverutils.ForceTableGC(t, s, db2, "test", "t", s.Clock().Now().Add(-int64(1*time.Second), 0))
+			sqlutils.ForceTableGC(
+				t, s.DistSenderI().(*kv.DistSender), db2,
+				"test", "t", s.Clock().Now().Add(-int64(1*time.Second), 0))
 			time.Sleep(10 * time.Millisecond)
 		}
 	}()
