+++ pkg/testutils/serverutils/test_server_shim.go
@@ -18,7 +18,6 @@
 package serverutils
 
 import (
-	"context"
 	gosql "database/sql"
 	"net/http"
 	"net/url"
@@ -26,8 +25,6 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/internal/client"
-	"github.com/cockroachdb/cockroach/pkg/keys"
-	"github.com/cockroachdb/cockroach/pkg/kv"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/rpc"
 	"github.com/cockroachdb/cockroach/pkg/security"
@@ -238,29 +235,4 @@ func PostJSONProto(ts TestServerInterface, path string, request, response protou
 		return err
 	}
 	return httputil.PostJSON(httpClient, ts.AdminURL()+path, request, response)
-}
-
-// ForceTableGC sends a GCRequest for the ranges corresponding to a table.
-func ForceTableGC(
-	t testing.TB,
-	ts TestServerInterface,
-	db sqlutils.DBHandle,
-	database, table string,
-	timestamp hlc.Timestamp,
-) {
-	t.Helper()
-	tblID := sqlutils.QueryTableID(t, db, database, table)
-	tblKey := roachpb.Key(keys.MakeTablePrefix(tblID))
-	gcr := roachpb.GCRequest{
-		RequestHeader: roachpb.RequestHeader{
-			Key:    tblKey,
-			EndKey: tblKey.PrefixEnd(),
-		},
-		Threshold: timestamp,
-	}
-	if _, err := client.SendWrapped(
-		context.Background(), ts.DistSenderI().(*kv.DistSender), &gcr,
-	); err != nil {
-		t.Error(err)
-	}
 }
